it-generator/requirements.txt
torch==1.7.1
torchvision==0.8.2
pyyaml==5.3.1
h5py==3.1.0
numpy==1.19.4
pandas==1.1.5
matplotlib==3.3.3
tqdm==4.54.1
tensorboard==2.4.0
pillow==8.0.1
transformers==4.1.1
opencv-python==4.4.0.46
scipy==1.1.0
tensorflow==1.14.0
git+https://github.com/openai/CLIP.git
git+https://github.com/salaniz/pycocoevalcap